+++ crates/swc/tests/fixture/issues-4xxx/4851/input/index.ts
@@ -1,0 +1,13 @@
+import { TSESTree, AST_TOKEN_TYPES } from '../ts-estree';
+
+namespace AST {
+    export type TokenType = AST_TOKEN_TYPES;
+
+    export type Token = TSESTree.Token;
+
+    export type SourceLocation = TSESTree.SourceLocation;
+
+    export type Range = TSESTree.Range;
+}
+
+export { AST };