+++ crates/swc/tests/fixture/issues-5xxx/5626/input/index.ts
@@ -1,0 +1,8 @@
+function Prop(): PropertyDecorator {
+    return function () { };
+}
+
+export default class Example {
+    @Prop()
+    prop: bigint;
+}