+++ crates/swc/tests/tsc-references/exportsAndImports1-amd.1.normal.js
@@ -39,9 +39,6 @@ define([
         M: function() {
             return M;
         },
-        N: function() {
-            return N;
-        },
         a: function() {
             return a;
         }