+++ crates/swc/tests/tsc-references/exportsAndImports1-es6.2.minified.js
@@ -16,7 +16,6 @@ Object.defineProperty(exports, "__esModule", {
     E: ()=>E,
     D: ()=>D,
     M: ()=>M,
-    N: ()=>N,
     a: ()=>a
 });
 var E, D, M, v = 1;