+++ crates/swc/tests/tsc-references/exportsAndImports3-es6.1.normal.js
@@ -25,7 +25,6 @@ _export(exports, {
     E1: ()=>E,
     D1: ()=>D,
     M1: ()=>M,
-    N1: ()=>N,
     a1: ()=>a
 });
 var v = 1;