+++ crates/swc/tests/tsc-references/exportsAndImports3.1.normal.js
@@ -51,9 +51,6 @@ _export(exports, {
     M1: function() {
         return M;
     },
-    N1: function() {
-        return N;
-    },
     a1: function() {
         return a;
     }