+++ CHANGELOG.md
@@ -10,6 +10,21 @@
 
 - **(es/minifier)** Don't inline variables declared multiple times (#5667) ([fd39bb6](https://github.com/swc-project/swc/commit/fd39bb6bff45d3200ad1934d4b2b470a7c8ec21c))
 
+
+- **(es/typescript)** Mark a namespace as concrete only if it is (#5670) ([4babf4a](https://github.com/swc-project/swc/commit/4babf4a2c0467926410bdcb45860a53f3ec59128))
+
+### Features
+
+
+
+- **(es/decorators)** Support `bigint` (#5671) ([b1f1a86](https://github.com/swc-project/swc/commit/b1f1a86a9df32e4bc0d9210d0a90ada8a0121cca))
+
+### Testing
+
+
+
+- **(es/compat)** Add a test for a fixed issue (#5672) ([1ae539a](https://github.com/swc-project/swc/commit/1ae539aac52cd25dec8be35799e9a56bb513a660))
+
 ## [1.2.245] - 2022-08-30
 
 ### Bug Fixes